requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
python-docx==1.0.1
lxml==4.9.3
Pillow==10.0.0
python-slugify==8.0.1
pytz==2023.3